+++ src/scripts/create_scala.py
@@ -13,6 +13,7 @@ from load_ud_pos import (
     load_dadt_pos,
     load_dedt_pos,
     load_endt_pos,
+    load_esdt_pos,
     load_fodt_pos,
     load_frdt_pos,
     load_isdt_pos,
@@ -45,6 +46,7 @@ def main() -> None:
         "nl": load_nldt_pos,
         "en": load_endt_pos,
         "fr": load_frdt_pos,
+        "es": load_esdt_pos,
     }
 
     # Set up the progress bar and iterate over the languages